+++ src/Tests/EventGrid.fs
@@ -46,19 +46,21 @@ let tests = testList "Event Grid" [
         Expect.equal sub.Destination hub.EventHubNamespaceName "Incorrect destination"
     }
     test "Creates a service bus queue subscriber correctly" {
-        let bus = serviceBus { name "busbus"; add_queues [ queue { name "queue" } ] }
-        let grid = eventGrid { add_servicebus_subscriber ServiceBusSubscriptionType.Queue  bus [] }
+        let q = queue { name "queuequeue" }
+        let bus = serviceBus { name "busbus"; add_queues [ q ] }
+        let grid = eventGrid { add_servicebus_queue_subscriber bus q [] }
         let sub = grid.Subscriptions.[0]
-        Expect.equal sub.Name (ResourceName "busbus-busbus-servicebus-queue") "Incorrect subscription name"
-        Expect.equal sub.Endpoint (EndpointType.ServiceBus (ServiceBusEndpointType.Queue bus.Name)) "Incorrect endpoint type"
-        Expect.equal sub.Destination bus.Name "Incorrect destination"
+        Expect.equal sub.Name (ResourceName "queuequeue-busbus-servicebus-queue") "Incorrect subscription name"
+        Expect.equal sub.Endpoint (EndpointType.ServiceBus (ServiceBusEndpointType.Queue { Queue = q.Name; Bus = bus.Name })) "Incorrect endpoint type"
+        Expect.equal sub.Destination q.Name "Incorrect destination"
     }
     test "Creates a service bus topic subscriber correctly" {
-        let bus = serviceBus { name "busbus"; add_queues [ queue { name "queue" } ] }
-        let grid = eventGrid { add_servicebus_subscriber ServiceBusSubscriptionType.Topic  bus [] }
+        let t = topic { name "topictopic" }
+        let bus = serviceBus { name "busbus"; add_topics [ t ] }
+        let grid = eventGrid { add_servicebus_topic_subscriber bus t [] }
         let sub = grid.Subscriptions.[0]
-        Expect.equal sub.Name (ResourceName "busbus-busbus-servicebus-topic") "Incorrect subscription name"
-        Expect.equal sub.Endpoint (EndpointType.ServiceBus (ServiceBusEndpointType.Topic bus.Name)) "Incorrect endpoint type"
-        Expect.equal sub.Destination bus.Name "Incorrect destination"
+        Expect.equal sub.Name (ResourceName "topictopic-busbus-servicebus-topic") "Incorrect subscription name"
+        Expect.equal sub.Endpoint (EndpointType.ServiceBus (ServiceBusEndpointType.Topic { Topic = t.Name; Bus = bus.Name })) "Incorrect endpoint type"
+        Expect.equal sub.Destination t.Name "Incorrect destination"
     }
 ]