+++ RELEASE_NOTES.md
@@ -6,8 +6,10 @@ Release Notes
 * Functions: Support for 64 bits.
 * Storage: Add support for tables
 * Event Grid: Ensure destination Queues are created as a dependency
+* Event Grid: Add ServiceBus Queue and Topic as supported destinations
 
 * Framework: Updated DeterministicGuid for RFC 4122 compatibility
+
 
 ## 1.4.0
 * Bing Search: Support for Bing Search (migrated from Cognitive Services).